Fix Unix builds
