+++ src/System.Private.CoreLib/shared/Interop/Windows/Advapi32/Interop.EVENT_INFO_CLASS.cs
@@ -9,11 +9,11 @@ internal partial class Interop
 {
     internal partial class Advapi32
     {
-        [DllImport(Libraries.Advapi32, ExactSpelling = true)]
-        internal static unsafe extern int EventSetInformation(
-            [In] long registrationHandle,
-            [In] EVENT_INFO_CLASS informationClass,
-            [In] void* eventInformation,
-            [In] int informationLength);
+        internal enum EVENT_INFO_CLASS
+        {
+            BinaryTrackInfo,
+            SetEnableAllKeywords,
+            SetTraits,
+        }
     }
 }
