+++ src/System.Private.CoreLib/shared/Interop/Windows/Advapi32/Interop.EventActivityIdControl.cs
@@ -9,15 +9,6 @@ internal partial class Interop
 {
     internal partial class Advapi32
     {
-        internal enum ActivityControl : uint
-        {
-            EVENT_ACTIVITY_CTRL_GET_ID = 1,
-            EVENT_ACTIVITY_CTRL_SET_ID = 2,
-            EVENT_ACTIVITY_CTRL_CREATE_ID = 3,
-            EVENT_ACTIVITY_CTRL_GET_SET_ID = 4,
-            EVENT_ACTIVITY_CTRL_CREATE_SET_ID = 5
-        };
-
         [DllImport(Libraries.Advapi32, ExactSpelling = true)]
         internal static extern int EventActivityIdControl([In] ActivityControl ControlCode, [In][Out] ref Guid ActivityId);
     }
